Replaced DnDLayout with Editor, using LayoutProvider
