--- src/Layout.js
@@ -1,7 +1,0 @@
-import React  from 'react';
-
-import { connect } from './utils';
-
-const Layout = () => (
-  
-);
